Finalizing Presentation
Finishing up the presentation and moving some files around. Might return to the Markdown format someday. It does mostly, it just did not want to show code and results for some reason on the slides. Need more time to debug it, but it's far easier to do that in R anyhow.